refactor: move highlighter to dom renderer
