+++ src/chart_types/partition_chart/renderer/dom/highlighter.tsx
